requirements.txt
bcrypt==4.1.3
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pycryptodome==3.20.0
SQLAlchemy==2.0.30
typing_extensions==4.11.0
Werkzeug==3.0.3
WTForms==3.1.2
